*AwtGameWindow cleanups and extractions
